requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
Pillow>=10.0.0
python-decouple==3.8
crispy-bootstrap5==0.7
django-crispy-forms==2.0

setuptools<81
